Core: Split project translation responsibilities
